SRVSServer.NetrShareGetInfo() fix when non existent share was sent.
The SMBServer has a mini DCERPC server serving a minimal subset of
[MS-SRVS] calls used by clients.
The NetrShareGetInfo() method was assuming the share asked for
information was always existing, something that could not be true.
Calling, from the client, 'start \\smbserver\SHARE', and SHARE being
non existant was triggering this problem.
Now if the share doesn't exist the method returns WEE_NET_NAME_NOT_FOUND.