Adding support for connecting to an OSX Yosemite SMB share.
Yosemite's SMB implementation would complain if the following flags are not
part of the protocol negotiation:
FLAGS2_NT_STATUS
FLAGS2_LONG_NAMES
FLAGS2_UNICODE

If the first nego packet does not have that, the connection is closed.
We don't really want to announce UNICODE support for all servers (especially
against a NT4/SMB1 server), so we're changing these flags only if the
former flags were rejected.