SMB3.create: define CreateContextsOffset and CreateContextsLength when applicable
When providing createContexts in a SMB3 create call, the buffer is set
accordingly, but not the offset and length fields. This causes the createContext
data to be ignored by the server.

This patch tries to fix this by specifying the offset and length. The offset is
determined by adding the size of the SMB2 and CREATE_REQUEST headers to the
name length.