Trying to detect the type of format offered when called with -just-dc-user
Depending on the input received, we need to change the formatOffered before calling
DRSCrackNames.
There are some instances when you call -just-dc-user and you receive ERROR_DS_NAME_ERROR_NOT_UNIQUE
That's because we don't specify the domain for the user (and there might be duplicates)
Always remember that if you specify a domain, you should specify the NetBIOS domain name,
not the FQDN. Just for this time. It's confusing I know, but that's how this API works

If you receive ERROR_DS_NAME_ERROR_NOT_UNIQUE try calling the script again with <NetBIOS Domain Name>/<username> format