Decode the CONTROL_PAGEDRESULTS class in order to extract the Cookie
With previous pyasn1 versions (e.g. 0.2.3), this was automatically done by means of
overridding the __class__ variable (check Control.setComponentByPosition()) when a Control object was decoded.
Although not the cleanest approach, it's been working so far.

With newer pyasn1 versions (e.g. 0.3.3) is it not possible to override the __class__ variable anymore,
so we need to come up with a solution when it comes the time of using a paged search control.
For now, I'm decoding the Control() instance into a KNOWN_CONTROLS[CONTROL_PAGEDRESULTS].
I say for now, because I'm not convinced of encoding/decoding again something that hopefully could've been solved
before in the execution path, but this is what I came up with. There's a performance penalty that I'd like to avoid.

Maybe @kacpern can improve this code ;)