Trying to increase the likelihood of getting RC4 session tickets.
In order to maximize the probability of getting session tickets with RC4 etype, we will convert the
password to ntlm hashes (that will force to use RC4 for the TGT). If that doesn't work, we use the
cleartext password.
If no clear text password is provided, we just go with the defaults.