Encrypting the TGS-REP with the right Key Usage (especially for AES)
* When encrypting the kdcRep['enc-part']['cipher'] part you have to use the following Key Usages depending on
the type of answer being generated. Per RFC 4120, section 7.5.1:
If it is a AS-REP -> Key Usage 3
If it is a TGS-REP -> Key Usage 8
When silver tickets functionality was implemented [here](https://github.com/CoreSecurity/impacket/pull/354),
this wasn't taken into account that all the answers, regardless of being a AS-REP or TGS-REP were encrypted with Key Usage 3.

You might wonder why RC4 silver tickets where working?
RFC4757 section 3 has the answer. The Key Usages for RC4 are different by the ones defined in RFC4120 and,
for example, for the AS-REP the Key Usage to be used is 8 (instead of 3). Thus, all the RC4 testings were
passing, until @ropnop tested it with AES keys (where the Key Usage numbers to be used are the ones in
RFC 4120), causing the decryption checks when saving the ticket in ccache format to trigger and error
due to invalid checksum.

The issue in https://github.com/CoreSecurity/impacket/issues/443#issuecomment-384430424 should be fixed. Please test.