Adding getST.py example
Given a password, hash, aesKey or TGT in ccache, it will request a Service Ticket
and save it as ccache.
If the account has constrained delegation (with protocol transition) privileges
you will be able to use the -impersonate switch to request the ticket on behalf
other user (it will use S4U2Self/S4U2Proxy to request the ticket.)

Similar feature has been implemented already by Benjamin Delphi (@gentilkiwi) in Kekeo (s4u)

Examples:

    ./getST.py -hashes lm:nt -spn cifs/contoso-dc contoso.com/user
It will save the ticket in user.ccache
or
If you have tickets cached (run klist to verify) the script will use them
    ./getST.py -k -spn cifs/contoso-dc contoso.com/user
Be sure tho, that the cached TGT has the forwardable flag set (klist -f).
getTGT.py will ask forwardable tickets by default.

Also, if the account is configured with unconstrained delegation (and with protocol transition)
you can request service tickets for other users, assuming the target SPN is allowed for
delegation:
./getST.py -k -impersonate Administrator -spn cifs/contoso-dc contoso.com/user

The output of this script will be a service ticket for the Administrator user.

Once you have the ccache file, set it in the KRB5CCNAME variable and use it for fun and profit.