Catch decryption error and add context information
When decrypting session key from the AS-REP message, we might
encounter decryption error such as invalid checksum if the
provided key is not the right one (when pre-auth is disabled
for instance).

We now add a SessionKeyDecryptionError exception with context
information about the decryption error (key, cipher, cipherText
and AsRep). This may give information to the exception handler
about what happened.

Finally, the exception handler will be able to compute john hash
from the exception, which was not possible with the InvalidChecksum
exception.