Adding DPAPI / Vaults related structures, testcase and example.
You can unlock masterkeys, credentials and vaults. For the three, you will specify the file name (using -file for
masterkeys and credentials, and -vpol and -vcrd for vaults).
If no other parameter is sent, the contents of these resource will be shown, with their encrypted data as well.
If you specify a -key blob (in the form of '0xabcdef...') that key will be used to decrypt the contents.
In the case of vaults, you might need to also provide the user's sid (and the user password will be asked).
For system secrets, instead of a password you will need to specify the system and security hives.

References: All of the work done by these guys. I just adapted their work to my needs.
https://www.passcape.com/index.php?section=docsys&cmd=details&id=28
https://github.com/jordanbtucker/dpapick
https://github.com/gentilkiwi/mimikatz/wiki/howto-~-credential-manager-saved-credentials (and everything else Ben did )
http://blog.digital-forensics.it/2016/01/windows-revaulting.html
https://www.passcape.com/windows_password_recovery_vault_explorer
https://www.passcape.com/windows_password_recovery_dpapi_master_key