Fix: default list shared across all Machine objects
In python, the default values of function parameters are instantiated
at function definition time. All calls to that function that use the
default value all point to the same global object. e.g.:

```
def func(x=[]):
   x.append(1)
   print(x)

func() # [1]
func() # [1 , 1]
```

Because of this, two instances of Machine (initialized without passing
in a protocol_config option) actually have self.openUDPPorts point to
the same list. So calling openUDPPort on one Machine object will also
possibly add that port to another Machine object.

Fix:
The recommended solution is to either set default to None and assign
a new empty object when the variable is None.