smbserver.py: check case insensitive NTLM username
Windows usernames are case insensitive. In smbserver credentials are
stored in the server in the exact manner they are provided by the user and fetched
using the username provided in the AUTH message as the key of the
dictionnary. If there is a difference in the case, credentials are not
found and the authentication fails.