smb3: store Connection.PreAuthIntegrityHashValue
Only the Session.PreAuthIntegrityHashValue is computed and stored.

As described in [MS-SMB2] a Connection.PreAuthIntegrityHashValue must be
computed and used to initialize the Session.PreAuthIntegrityHashValue
for all successive sessions.
This is important in the case of the SMBv3.1.1 dialect as it will allow
to obtain different signing keys for each session.