ldapattack: fix error when trying to escalate with machine account
When relaying to ldap/ldaps and trying to escalate with a machine
account, the machine account is found by getUserInfo() but not by
aclAttack(). aclAttack assumes that the objectCategory=user which
machine accounts are not. By being a little more generic,
objectClass=user will allow to use both categories of objects.