Changing -auth-rpcproxy function to -auth-transport in rpcmap.py and modifying the code to suit the new MS-RPCH impl
+ Now it's possible to specify different credentials for SMB and RPC
+ RPC Proxy errors are handled by constants from its implementation

The downside of this commit is it removes Kerberos support.

Current SMB and RPC libraries depend too much on each other,
we need to make some changes for them to get an ability to use different credentials and
different security providers on both levels, and add a support to make reconenctions with
proper socket handeling, and saving security configuration.

Probably removing Kerberos is not a big deal, as rpcmap.py is usually used in research purposes.