httpRelayServer: read body content for POST requests
Web clients receiving a 401 Unauthorized header as a response, depending
on the Keep-alive header, may authenticate inside the same TCP stream,
or close the first one and open another.
In the first case, if the client was sending a POST request with data in
the content body, the content body was parsed by handle_one_request()
as a new (invalid) request, thus causing an issue with the parsing of the following
request where the Authorization header was actually presented.