Change duration measurement from days to hours
The default value for the lifetime of a legit ticket is 10h, so a longer one could be suspicious.
It is not possible to set a lifetime shorter than 1 day due to the fact that the data type of the duration parameter is int and it is measured in days.

The measurement is changed from days to hours because the value considered opsec is 10h. The default value for the duration parameter is still 10 years for anyone who doesn't care about opsec. The data type is now float in case someone wants to set a decimal value.