Added basic GitHub Actions workflow (#1055)
As we were running with some issues with Travis (e.g. limited run minutes), we're adding GitHub Actions as a separate CI unit tests runner. Tests are executed via Tox in virtual environments so we should expect the same results as in Travis, although we're not yet replacing it at this point.

* Adds a basic workflow that:
  * Runs flake8 for syntax errors and then runs non-remote unit tests using Tox.
  * Runs on Python 2.7, 3.6, 3.7, 3.8 and allows failures on 3.9 as we do in Travis.
* Adds a job that just builds the contributed docker image to check it's not failing.