ticketer: Add support for MIT Kerberos V
TGTs indicate the kvno of the key used to sign them. Microsoft's
implementation of Kerberos 5 does not verify that the kvno specified
matches an existing key. MIT Kerberos V, on the other hand, does.
The use of a hard-coded kvno (with value 2) causes MIT's implementation
to refuse tickets signed by ticketer.py

This patch adds support for the `-kvno` argument. It lets users
specify the kvno of the signing key. Additionally, since a domain's
SID only exists in Active Directory, the argument becomes optional.

At least 1 of the argument must be specified. If the kvno is not specified,
the currently hard-coded value is used as the default. If the domain SID
is not specified, a default value is used. Note that, to prevent users from
specifying a kvno when targeting an AD, a note about only using the kvno
argument for targeting MIT is given.

For reference, do note Microsoft did document that it ignores the kvno field:
https://docs.microsoft.com/en-us/archive/blogs/openspecification/to-kvno-or-not-to-kvno-what-is-the-version