Allow search domain to be specified in GetADUsers
So far, GetADUsers.py only allowed searching for users within the same
domain. When running against a foreign domain, that trusts the domain
the current user is authenticating with, no useres were returned. This
commit adds the -search-domain option that allows to specify the domain
from which users should be enumerated.