dpapi: Fix printing master key
The code looks for leading zero bytes and removes them, but then uses
the number of zero bytes (in variable beginning) to trim the key again
just before printing. This commit fixes the code to only trim once.

The code was verified to work and print the same DPAPI master key as
Mimikatz and the other methods implemented by Impacket.