Added Cross-Domain Trust Capabilities
Fixed a logic bug that originally retrieved the LDAP domain search scope from the specified user credentials. 
A new flag now exists called -targetdomain, which can be used to ASREP-Roast in a targeted domain of your choice. 

The Argument Parsing function was modified to allow for this new functionality as well as the getTGT function to check if a target domain is specified. If so, it will use that instead of the domain. Lastly, the init function for declaring the LDAP search scope.