Fix version parsing in git
In `setup.py`, the date and time of the last commit is determined using
`git log`, but fails to take into account that users may have enabled
`log.showSignature` in their global git config. If this option is true
and a commit is signed, it breaks the parsing of the git output. As a
result, the package fails to build, because the version number is
invalid.

By just using the last line of the git output, we ensure that the
signature is ignored.